added todo notes on known bug
TODO: while working on this I realised that the script that the
cron job would call is the one that's written by the last set
cron job so say for example I set a recurring alarm for 8am
that played the CBC then set another for 10am that played
ABC when the 8am cron job called the script ABC would play.
I'll need a way to diferentiate the script that cron calls
to execute the alarm.

There's a reason this is called the testing branch.