Added identifier to shell script name
Added the value of date and time to the name
of the generated alarm shell script.  As long as no
one sets an alarm to happen at the same date and time
as another alarm (which would result in the radio
playing two streams at once) I think this will work
to fix the bug where recurring alarms only play the
stream of the last set alarm.

Also removed a TODO comment that was done regarding
crontab not appending new jobs.