bugfix
non-recurring alarm metadata was not being removed from the db
after the alarm went off (i.e., the at job ran) this was because
the value of time stored in the DB included ':00' seconds while
the value of the time variable only contained hours and minutes.

Also removed debugging output.